# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TokenTimer Core API Contract
  version: 0.1.0
  description: |
    TokenTimer Core API for authentication, workspaces, token lifecycle management,
    alerts, audit, and integration import/scan workflows.
    This contract is also used as a compatibility baseline across core, cloud, and enterprise variants.
servers:
  - url: http://localhost:4000
    description: Local development backend
...
security:
  - cookieAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: loginWithPassword
      description: Authenticates a user and returns either a session payload or a 2FA challenge.
      security: []
      requestBody:
        description: Credentials used for password authentication.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success or 2FA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/LoginSuccessResponse"
                  - $ref: "#/components/schemas/LoginNeeds2FAResponse"
        "401":
          description: Invalid credentials or unverified account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session cookie authentication used by authenticated API routes.
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          oneOf:
            - type: array
              items:
                type: string
            - type: object
              additionalProperties: true
            - type: string
      example:
        error: Invalid request payload
        code: VALIDATION_ERROR
        details:
          - email is required
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@localhost.local
        password:
          type: string
          example: AdminPassword123!
    LoginSuccessResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            email:
              type: string
            displayName:
              type: string
    LoginNeeds2FAResponse:
      type: object
      required: [requires2FA]
      properties:
        requires2FA:
          type: boolean
        authMethod:
          type: string
...